# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WH40k Colony Manager API",
    "description": "REST API for managing Warhammer 40k Rogue Trader colonies",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/colonies/{colony_id}/members": {
      "get": {
        "tags": [
          "colony_users"
        ],
        "summary": "Get Colony Members",
        "description": "Get all members of a colony with pagination.\n\nNote: Pagination is applied in-memory after loading all items. This is acceptable\nfor typical colony sizes (<100 members). For colonies with >1000 members,\nconsider adding paginated query methods to the repository layer.",
        "operationId": "get_colony_members_api_v1_colonies__colony_id__members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ColonyUserListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "colony_users"
        ],
        "summary": "Add Colony Member",
        "description": "Add a user to a colony.\n\nRequires appropriate permissions (typically owner or editor role).",
        "operationId": "add_colony_member_api_v1_colonies__colony_id__members_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColonyUserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColonyUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/colonies/{colony_id}/members/{user_id}": {
      "get": {
        "tags": [
          "colony_users"
        ],
        "summary": "Get Colony Member",
        "description": "Get a specific user's membership in a colony.",
        "operationId": "get_colony_member_api_v1_colonies__colony_id__members__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColonyUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "colony_users"
        ],
        "summary": "Update Colony Member Role",
        "description": "Update a member's role in the colony.\n\nRequires appropriate permissions (typically owner role).",
        "operationId": "update_colony_member_role_api_v1_colonies__colony_id__members__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColonyUserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColonyUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "colony_users"
        ],
        "summary": "Remove Colony Member",
        "description": "Remove a user from a colony.\n\nRequires appropriate permissions (typically owner role).",
        "operationId": "remove_colony_member_api_v1_colonies__colony_id__members__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/colonies/{colony_id}/members/transfer-ownership": {
      "post": {
        "tags": [
          "colony_users"
        ],
        "summary": "Transfer Colony Ownership",
        "description": "Transfer colony ownership to another user.\n\nRequires owner-level permissions. The current owner is demoted to editor\nby default (can be disabled with demote_current=False).\n\nReturns:\n    Dictionary with new_owner and previous_owner user IDs.",
        "operationId": "transfer_colony_ownership_api_v1_colonies__colony_id__members_transfer_ownership_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "colony_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Colony Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColonyOwnershipTransfer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Transfer Colony Ownership Api V1 Colonies  Colony Id  Members Transfer Ownership Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "description": "List all users with pagination.\n\nRequires admin privileges. Returns paginated list of users.\n\nRaises:\n    HTTPException: 403 if user is not an admin.",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of users to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of users to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of users to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of users to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UserListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Create a new user.\n\nRequires admin privileges. Creates a new user account with the provided details.\nUsername and email are immutable after creation.\n\nRaises:\n    HTTPException: 403 if user is not an admin.\n    HTTPException: 400 if username or email already exists.",
        "operationId": "create_user_api_v1_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/colony_manager__adapters__api__schemas__user__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "description": "Get a specific user by ID.\n\nRequires admin privileges.\n\nRaises:\n    HTTPException: 403 if user is not an admin.\n    HTTPException: 404 if user not found.",
        "operationId": "get_user_api_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/colony_manager__adapters__api__schemas__user__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update User",
        "description": "Update a user.\n\nRequires admin privileges. Username and email are immutable.\nAdmins cannot modify other admins, and users cannot escalate their own privileges.\n\nRaises:\n    HTTPException: 403 if user is not an admin or lacks permission.\n    HTTPException: 404 if user not found.\n    HTTPException: 400 if validation fails.",
        "operationId": "update_user_api_v1_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/colony_manager__adapters__api__schemas__user__UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User",
        "description": "Delete a user (soft delete).\n\nRequires admin privileges. Sets is_active to False instead of hard deleting.\nAdmins cannot delete other admins.\n\nRaises:\n    HTTPException: 403 if user is not an admin or lacks permission.\n    HTTPException: 404 if user not found.",
        "operationId": "delete_user_api_v1_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColonyOwnershipTransfer": {
        "properties": {
          "new_owner_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "New Owner Id",
            "description": "User ID of the new owner"
          },
          "demote_current": {
            "type": "boolean",
            "title": "Demote Current",
            "description": "Whether to demote current owner to editor",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "new_owner_id"
        ],
        "title": "ColonyOwnershipTransfer",
        "description": "Request schema for transferring colony ownership."
      },
...
      "ColonyUserCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/ColonyUserRoleEnum",
            "default": "viewer"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "ColonyUserCreate",
        "description": "Schema for creating a colony-user membership."
      },
...
      "ColonyUserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "colony_id": {
            "type": "integer",
            "title": "Colony Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          },
          "invited_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invited By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "colony_id",
          "user_id",
          "role",
          "joined_at",
          "invited_by"
        ],
        "title": "ColonyUserResponse",
        "description": "Schema for colony-user membership response."
      },
...
      "ColonyUserUpdate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/ColonyUserRoleEnum"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "ColonyUserUpdate",
        "description": "Schema for updating a colony-user membership."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ColonyUserListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ColonyUserListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse[ColonyUserListItem]"
      },
...
      "PaginatedResponse_UserListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "PaginatedResponse[UserListItem]"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "description": "Username (immutable after creation)"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address (immutable after creation)"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Initial password"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "User role: viewer, colony_manager, or admin",
            "default": "viewer"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the user account is active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "UserCreate",
        "description": "Request schema for creating a new user (admin only)."
      },
...
      "colony_manager__adapters__api__schemas__user__UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "role",
          "is_active"
        ],
        "title": "UserResponse",
        "description": "Response schema for user information (excludes sensitive data)."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (do not include 'Bearer' prefix in the value)"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}